NF: Refactor alert handling: Move SweetAlert utility functions to a separate file for better organization and reuse. Update components to utilize the new alert functions.
