+++ frontend/src/components/Pages/AdminPanal.js
@@ -4,8 +4,11 @@ import { useUI } from '../../context/UIContext';
 import { getProfessors, assignCourse, unassignCourse } from '../../services/professorService';
 import Sidebar from '../Common/Sidebar';
 import ToggleButton from '../ui/ToggleButton';
-import Swal from 'sweetalert2';
-import withReactContent from 'sweetalert2-react-content';
+import { 
+  showSuccessAlert, 
+  showErrorAlert, 
+  showConfirmDialog 
+} from '../../utils/sweetAlert';
 import {
   Plus,
   Search,
@@ -20,55 +23,6 @@ import {
   Edit2,
   Trash2,
 } from 'lucide-react';
-
-// Initialize SweetAlert2
-const MySwal = withReactContent(Swal);
-
-// Utility functions for alerts
-const showSuccessAlert = (title, text) => {
-  return MySwal.fire({
-    title,
-    text,
-    icon: 'success',
-    toast: true,
-    position: 'top-end',
-    showConfirmButton: false,
-    timer: 3000,
-    timerProgressBar: true,
-    background: document.documentElement.classList.contains('dark') ? '#2A2A2A' : '#fff',
-    color: document.documentElement.classList.contains('dark') ? '#fff' : '#000',
-  });
-};
-
-const showErrorAlert = (title, text) => {
-  return MySwal.fire({
-    title,
-    text,
-    icon: 'error',
-    toast: true,
-    position: 'top-end',
-    showConfirmButton: false,
-    timer: 3000,
-    timerProgressBar: true,
-    background: document.documentElement.classList.contains('dark') ? '#2A2A2A' : '#fff',
-    color: document.documentElement.classList.contains('dark') ? '#fff' : '#000',
-  });
-};
-
-const showConfirmDialog = (title, text, confirmButtonText = 'Yes', cancelButtonText = 'No') => {
-  return MySwal.fire({
-    title,
-    text,
-    icon: 'warning',
-    showCancelButton: true,
-    confirmButtonText,
-    cancelButtonText,
-    background: document.documentElement.classList.contains('dark') ? '#2A2A2A' : '#fff',
-    color: document.documentElement.classList.contains('dark') ? '#fff' : '#000',
-    confirmButtonColor: '#3B82F6',
-    cancelButtonColor: '#EF4444',
-  });
-};
 
 const AdminCourseManagement = () => {
   const { user } = useAuth();
